api/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoicer_api"
description = "This is a simple reference Python application using pyproject.toml."
classifiers = ["Private :: Do Not Upload"]
version = "0.0.1"
dependencies = [
    "requests==2.28.1",
    "Flask==2.2.2",
    "fpdf2==2.6.0",
    "Authlib==1.2.0"
]
requires-python = ">=3.11"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest"
]

